Add files via upload
In the face of the unprecedented COVID-19 pandemic, our data exploration project aims to provide a deep and insightful analysis of the global impact of the virus. Leveraging diverse datasets related to COVID-19, our project delves into various aspects, including infection rates, mortality trends, vaccination progress, and the socioeconomic implications of the pandemic.